validaciones al eliminar registros

+++ app/Http/Controllers/PaisController.php
@@ -4,9 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Pais;
-use App\Http\Requests\storePais;
-use App\Http\Requests\StorePais as RequestsStorePais;
 use Symfony\Contracts\Service\Attribute\Required;
+use Illuminate\Support\Facades\DB;
 
 class PaisController extends Controller
 {
@@ -15,7 +14,7 @@ class PaisController extends Controller
     {
         $this->middleware('auth');
     }
-    
+
     public function index()
     {
         $paises = Pais::All();
@@ -25,11 +24,11 @@ class PaisController extends Controller
 
     public function store(Request $request)
     {
-        
+
 
         $codigopais = $request->codigo;
         $nombrepais = $request->nombre;
-        
+
         $pais = new Pais();
         $pais->codigo = $codigopais;
         $pais->nombre = $nombrepais;
@@ -47,7 +46,7 @@ class PaisController extends Controller
         return view('pais.edit', $data);
     }
 
- 
+
     public function update(Request $request, $id)
     {
         $codigopais = $request->codigo;
@@ -56,16 +55,27 @@ class PaisController extends Controller
         $pais = Pais::findOrFail($id);
         $pais->codigo = $codigopais;
         $pais->nombre = $nombrepais;
-        
+
         $pais->update();
-        
+
         return redirect()->route('pais.index')->with('editado', 'ok');
     }
 
-    public function destroy($id)
-    {
-        $pais = Pais::findOrFail($id);
-        $pais->delete();
-        return redirect()->route('pais.index')->with('eliminado', 'ok');
+    public function destroy($id){
+
+        //Si el país está relacionado a un departamento no se puede eliminar.
+        $sql = 'SELECT  id_pais from departamentos
+        where id_pais = '.$id;
+
+        $departamentos_relacionados = DB::select($sql);
+
+        if(empty($departamentos_relacionados)){
+            $pais = Pais::findOrFail($id);
+            $pais->delete();
+            return redirect()->route('pais.index')->with('eliminado', 'ok');
+        }else{
+            return redirect()->route('pais.index')->with('error', 'ok');
+        }
+
     }
 }
